Changing default IPv4 MTU to 1260
If the client is IPv6-only and the server is IPv4-only, the traffic goes via NAT64. When IPv4 packets are translated to IPv6 ones, the header becomes 20 bytes longer. So it's better to use 1260 bytes MTU, to ensure that translated packets never exceed 1280 of min IPv6 MTU and do not get fragmented by NAT64.